+++ src/domain/mod.rs
@@ -15,7 +15,6 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-pub(crate) mod core;
 pub(crate) mod index;
 pub(crate) mod location;
 pub(crate) mod niseci;