+++ src/engines/niseci/x1/mod.rs
@@ -17,7 +17,7 @@
 
 use std::collections::HashMap;
 
-use crate::model::niseci::{CampionamentoNISECI, RecordNISECI, RiferimentoNISECI, SpecieNISECI};
+use crate::domain::niseci::{CampionamentoNISECI, RecordNISECI, RiferimentoNISECI, SpecieNISECI};
 
 pub(crate) fn calculate_x1(campionamento: &CampionamentoNISECI, riferimento: &RiferimentoNISECI) -> f32 {
   // n_i è il numero di specie autoctone di maggiore importanza ecologico-funzionale campionate