+++ src/views/mod.rs
@@ -22,16 +22,10 @@ use raylib::prelude::*;
 use raylib::consts::GuiState::{STATE_NORMAL, STATE_DISABLED};
 use raylib::consts::GuiIconName::{ICON_FILE_OPEN, ICON_BIN, ICON_OK_TICK, ICON_CROSS, ICON_PLAYER_NEXT};
 use rfd::FileDialog;
-use crate::model::core::SubModel;
-use crate::model::index::Indice;
-use crate::model::location::Location;
-use crate::model::niseci::TipoComunitaNISECI;
-use crate::model::niseci::ComunitaNISECI;
-use crate::model::niseci::AreaNISECI;
-use crate::model::niseci::IdroEcoRegioneNISECI;
-use crate::model::niseci::AnagraficaNISECI;
+use crate::domain::{index::Indice, location::Location, niseci::{TipoComunitaNISECI, ComunitaNISECI, AreaNISECI, IdroEcoRegioneNISECI, AnagraficaNISECI}};
 use crate::controllers::{Controller, Controllers, HomeController, SecondController, IndiceController, FileInputController, InfoAggiuntiveController, OutputController, ConsoleController};
 use crate::app::core::{CurrentView, MainState, propwidth, propheight, CONSOLE_FONT_DATA};
+use crate::app::model::SubModel;
 use crate::core::SHORT_PROJECT_VERSION;
 
 pub(crate) struct Views {